Extract classes to show prompt.
